A function must check the required dependencies of R models. Issue #29
A function must check the required dependencies of R models.
The function was updated to cover the following situations: no
dependencies, dependencies which are updated, lack of some dependencie.
A warning is presented if versions are different from installed package
and that required by the model, thus, the user has to decide what he
will do.
Fixes #29